+++ base/notification/notification.go
@@ -68,7 +68,7 @@ type Notification struct {
 }
 
 type Advisory struct {
-	AdvisoryID   int    `json:"advisory_id"`
+	AdvisoryID   int64  `json:"advisory_id"`
 	AdvisoryName string `json:"advisory_name"`
 	AdvisoryType string `json:"advisory_type"`
 	Synopsis     string `json:"synopsis"`
